Revert "[Docs] Swagger 작성 - 정주연"
