+++ src/main/java/com/kuit/findyou/domain/auth/repository/UserRepository.java
@@ -1,6 +1,6 @@
 package com.kuit.findyou.domain.auth.repository;
 
-import com.kuit.findyou.domain.user.model.User;
+import com.kuit.findyou.domain.auth.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
