+++ src/main/java/com/kuit/findyou/domain/report/exception/FileStorageException.java
@@ -1,4 +1,4 @@
-package com.kuit.findyou.domain.image.exception;
+package com.kuit.findyou.domain.report.exception;
 
 import com.kuit.findyou.global.common.response.status.ResponseStatus;
 import lombok.Getter;
