+++ src/main/java/com/kuit/findyou/domain/report/exception_handler/ControllerAdvice.java
@@ -1,7 +1,6 @@
 package com.kuit.findyou.domain.report.exception_handler;
 
-import com.kuit.findyou.domain.image.exception.FileStorageException;
-import com.kuit.findyou.domain.report.exception.ReportCreationException;
+import com.kuit.findyou.domain.report.exception.FileStorageException;
 import com.kuit.findyou.global.common.response.BaseErrorResponse;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
@@ -12,12 +11,13 @@ import org.springframework.web.bind.annotation.RestControllerAdvice;
 import static com.kuit.findyou.global.common.response.status.BaseExceptionResponseStatus.*;
 
 @Slf4j
-@RestControllerAdvice(basePackages = "com.kuit.findyou.domain.report")
+@RestControllerAdvice
 public class ControllerAdvice {
-    @ResponseStatus(HttpStatus.BAD_REQUEST)
-    @ExceptionHandler(ReportCreationException.class)
-    public BaseErrorResponse handle_ReportCreationException(ReportCreationException e) {
-        log.error("[handle_ReportCreationException]", e);
-        return new BaseErrorResponse(BAD_REQUEST, e.getMessage());
+    // 파일 업로드 실패
+    @ResponseStatus(HttpStatus.INTERNAL_SERVER_ERROR)
+    @ExceptionHandler(FileStorageException.class)
+    public BaseErrorResponse handle_FileStorageException(FileStorageException e) {
+        log.error("[handle_FileStorageException]", e);
+        return new BaseErrorResponse(INTERNAL_SERVER_ERROR, e.getMessage());
     }
 }
