+++ src/main/java/com/kuit/findyou/domain/report/repository/ImageRepository.java
@@ -1,6 +1,6 @@
-package com.kuit.findyou.domain.image.repository;
+package com.kuit.findyou.domain.report.repository;
 
-import com.kuit.findyou.domain.image.model.Image;
+import com.kuit.findyou.domain.report.model.Image;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
