+++ src/main/java/com/kuit/findyou/domain/report/repository/ReportRepository.java
@@ -1,6 +1,6 @@
 package com.kuit.findyou.domain.report.repository;
 
-import com.kuit.findyou.domain.user.model.User;
+import com.kuit.findyou.domain.auth.model.User;
 import com.kuit.findyou.domain.report.model.Report;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Pageable;
