+++ src/main/java/com/kuit/findyou/domain/report/repository/ViewedProtectingReportRepository.java
@@ -1,6 +1,6 @@
 package com.kuit.findyou.domain.report.repository;
 
-import com.kuit.findyou.domain.user.model.User;
+import com.kuit.findyou.domain.auth.model.User;
 import com.kuit.findyou.domain.report.model.ProtectingReport;
 import com.kuit.findyou.domain.report.model.ViewedProtectingReport;
 import org.springframework.data.domain.Pageable;
