+++ src/main/java/com/kuit/findyou/domain/report/repository/ViewedReportRepository.java
@@ -1,7 +1,9 @@
 package com.kuit.findyou.domain.report.repository;
 
-import com.kuit.findyou.domain.user.model.User;
+import com.kuit.findyou.domain.auth.model.User;
+import com.kuit.findyou.domain.report.model.ProtectingReport;
 import com.kuit.findyou.domain.report.model.Report;
+import com.kuit.findyou.domain.report.model.ViewedProtectingReport;
 import com.kuit.findyou.domain.report.model.ViewedReport;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.domain.Slice;
