+++ src/main/java/com/kuit/findyou/domain/report/service/ReportAnimalRetrieveService.java
@@ -1,6 +1,6 @@
 package com.kuit.findyou.domain.report.service;
 
-import com.kuit.findyou.domain.user.model.User;
+import com.kuit.findyou.domain.auth.model.User;
 import com.kuit.findyou.domain.auth.repository.UserRepository;
 import com.kuit.findyou.domain.report.dto.Card;
 import com.kuit.findyou.domain.report.dto.ReportCardDTO;
