+++ src/test/java/com/kuit/findyou/domain/report/repository/ReportRepositoryTest.java
@@ -1,10 +1,8 @@
 package com.kuit.findyou.domain.report.repository;
 
-import com.kuit.findyou.domain.user.model.User;
+import com.kuit.findyou.domain.auth.model.User;
 import com.kuit.findyou.domain.auth.repository.UserRepository;
 import com.kuit.findyou.domain.home.dto.ReportTag;
-import com.kuit.findyou.domain.image.model.Image;
-import com.kuit.findyou.domain.image.repository.ImageRepository;
 import com.kuit.findyou.domain.report.model.*;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
