+++ src/test/java/com/kuit/findyou/domain/report/service/AnimalRetrieveServiceTest.java
@@ -1,10 +1,8 @@
 package com.kuit.findyou.domain.report.service;
 
-import com.kuit.findyou.domain.user.model.User;
+import com.kuit.findyou.domain.auth.model.User;
 import com.kuit.findyou.domain.auth.repository.UserRepository;
 import com.kuit.findyou.domain.home.dto.ReportTag;
-import com.kuit.findyou.domain.image.model.Image;
-import com.kuit.findyou.domain.image.repository.ImageRepository;
 import com.kuit.findyou.domain.report.dto.Card;
 import com.kuit.findyou.domain.report.dto.TotalCardDTO;
 import com.kuit.findyou.domain.report.model.*;
@@ -37,8 +35,7 @@ class AnimalRetrieveServiceTest {
     @Autowired AnimalFeatureRepository animalFeatureRepository;
     @Autowired InterestReportRepository interestReportRepository;
     @Autowired BreedRepository breedRepository;
-    @Autowired
-    ImageRepository imageRepository;
+    @Autowired ImageRepository imageRepository;
 
     @Autowired ReportAnimalRetrieveService reportAnimalRetrieveService;
