+++ src/test/java/com/kuit/findyou/domain/report/service/ImageServiceTest.java
@@ -1,7 +1,6 @@
 package com.kuit.findyou.domain.report.service;
-import com.kuit.findyou.domain.image.model.Image;
-import com.kuit.findyou.domain.image.repository.ImageRepository;
-import com.kuit.findyou.domain.image.service.ImageService;
+import com.kuit.findyou.domain.report.model.Image;
+import com.kuit.findyou.domain.report.repository.ImageRepository;
 import org.apache.commons.io.IOUtils;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
